use i64 in tree

+++ library/cpp/containers/disjoint_interval_tree/disjoint_interval_tree.h
@@ -115,7 +115,7 @@ public:
             if (containingBegin->first < begin && begin < containingBegin->second) { // Contains begin.
                 if (containingBegin->second > end) { // Contains end.
                     const T prevEnd = containingBegin->second;
-                    Y_ASSERT(containingBegin->second - begin <= NumElements);
+                    Y_ASSERT(static_cast<size_t>(containingBegin->second - begin) <= NumElements);
 
                     Y_ASSERT(containingBegin->second - containingBegin->first > end - begin);
                     containingBegin->second = begin;
